requirements.txt
openai==0.27.4
python-dotenv==1.0.0
pyfiglet==0.8.post1
typer==0.7.0
langchain==0.0.133
google-search-results==2.4.2
shellingham==1.5.0.post1
python-statemachine==2.0.0
pydot==1.4.2
networkx==3.1
matplotlib==3.7.1
tiktoken==0.3.3
